config: move gangway parts to separate file
